+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 1.0.0-beta.24
 ### Bugfix
 * add (`dfls` mnemonics for floating-point exponent)
+* fix symbol syntax extensions (e.g. empty string interpolation `#""`) failing
+  with an unterminated expression error
 
 ## 1.0.0-beta.23
 ### Breaking